tests/wpt/tests/container-timing/tentative/ignore-with-containertiming-and-child-img.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Container Timing: an image inside a containertiming inside a containertiming-ignore</title>
<body>
<style>
body {
  margin: 0;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/container-timing/resources/container-timing-helpers.js"></script>
<script src="/element-timing/resources/element-timing-helpers.js"></script>
<script>
  let beforeRender;
  async_test(function (t) {
    assert_implements(window.PerformanceContainerTiming, "PerformanceContainerTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkContainerEntry(entry, 'div2_ct', 'img_id', beforeRender)
        checkRect(entry, [0, 100, 0, 100])
        checkContainerSize(entry, 10000);
      })
    );
    observer.observe({entryTypes: ['container']});

    // Add a div that is the container timing root
    const div1 = document.createElement('div');
    div1.setAttribute('containertiming-ignore', '');
    document.body.appendChild(div1);

    // Intermediate ignore should block paint events
    const div2 = document.createElement('div');
    div2.setAttribute('containertiming', 'div2_ct');
    div1.appendChild(div2)

    // Add image of width equal to 100 and height equal to 100.
    const img = document.createElement('img');
    img.src = '/container-timing/resources/square100.png';
    img.setAttribute('id', 'img_id');
    div2.appendChild(img);

    beforeRender = performance.now();
}, 'An image inside a containertiming, inside a containertiming-ignore is still reported.');
</script>

</body>
